Make WX() int so it can go negative.
WXPlus7 (the reg) was being set to 6, which this then converted to 255, which was interpreted as WX at the end of the line (e.g. hidden).

TESTED=DMG_ACID (eyeball) + Zelda now has the equipment area at bottom.

+++ include/ppu.h
@@ -110,7 +110,7 @@ class PPU {
   uint8_t GetWXPlus7();
 
   // WX with the adjustment applied since wx = 7 is pixel 0.
-  uint8_t WXPixelX() { return GetWXPlus7() - 7; }
+  int WXPixelX() { return GetWXPlus7() - 7; }
 
   uint8_t bgp();
   uint8_t obp0();
